# OpenApi/ToContainer.yml
openapi: 3.0.0
info:
  title: kms-api-gateway
  version: 1.0.0
...
servers:
  - url: https://d5dmup0d42c7f5ukd595.apigw.yandexcloud.net
paths:
...
  /auth/registration:
    post:
      x-yc-apigateway-integration:
        type: serverless_containers
        container_id: bbafcpeckkgctgerjgc2
      operationId: registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Пользователь создан
...
  /history/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: path
    get:
      x-yc-apigateway-integration:
        type: serverless_containers
        container_id: bbafcpeckkgctgerjgc2
        security:
          -jwtHeaderAuthorizer:
            - profile:read
            - profile:write
      operationId: view
      responses:
        '200':
          description: view
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/EditDto'
    post:
      x-yc-apigateway-integration:
        type: serverless_containers
        container_id: bbafcpeckkgctgerjgc2
        security:
          -jwtHeaderAuthorizer:
            - profile:read
            - profile:write
      operationId: saveVersion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVersionDto'
      responses:
        '200':
          description: saveVersion
  /history/versions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: path
    get:
      x-yc-apigateway-integration:
        type: serverless_containers
        container_id: bbafcpeckkgctgerjgc2
        security:
          -jwtHeaderAuthorizer:
            - profile:read
            - profile:write
      operationId: oldVersion
      responses:
        '200':
          description: oldVersion
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/TextDto'
    delete:
      x-yc-apigateway-integration:
        type: serverless_containers
        container_id: bbafcpeckkgctgerjgc2
        security:
          -jwtHeaderAuthorizer:
            - profile:read
            - profile:write
      operationId: deleteVersion
      responses:
        '200':
          description: deleteVersion
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/HistoryArticle'
  /panel/users:
    get:
      x-yc-apigateway-integration:
        type: serverless_containers
        container_id: bbafcpeckkgctgerjgc2
        security:
          -jwtHeaderAuthorizer:
            - profile:read
            - profile:write
      operationId: viewEmp
      responses:
        '200':
          description: viewEmp
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeAndRole'
    put:
      x-yc-apigateway-integration:
        type: serverless_containers
        container_id: bbafcpeckkgctgerjgc2
        security:
          -jwtHeaderAuthorizer:
            - profile:read
            - profile:write
      operationId: updateEmp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/User'
      responses:
        '200':
          description: updateEmp
components:
  schemas:
    User:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
        surname:
          type: string
        patronymic:
          type: string
        email:
          type: string
...
    EditDto:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        id_version:
          type: integer
        versions:
          type: array
          items:
            $ref: '#/components/schemas/VersionDto'
    VersionDto:
      type: object
      properties:
        id_version:
          type: integer
        createdAt:
          type: string
    UpdateVersionDto:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        id_article:
          type: integer
        id_version:
          type: integer
    TextDto:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
    HistoryArticle:
      type: object
      properties:
        id:
          type: integer
        id_article:
          type: integer
        title:
          type: string
        content:
          type: string
        createdAt:
          type: string
    EmployeeAndRole:
      type: object
      properties:
        name:
          type: string
        surname:
          type: string
        patronymic:
          type: string
        regDate:
          type: string
          format: date
        role:
          type: string
  securitySchemes:
    jwtHeaderAuthorizer:
      type: openIdConnect
      x-yc-apigateway-authorizer:
        type: jwt
        jwksUri: https://example.com/.well-known/jwks.json
        identitySource:
          in: header
          name: Authorization
          prefix: "Bearer "